refactor: Split the wasm simulator and recorder
