+++ apps/web-tools/package.json
@@ -26,6 +26,7 @@
     "react-markdown": "^10.1.0",
     "react-router": "^7.13.2",
     "remark-gfm": "^4.0.1",
+    "sim-recorder": "link:../../firmware/sim-recorder/pkg",
     "sim-wasm": "link:../../firmware/sim-wasm/pkg",
     "three": "^0.183.2",
     "vaul": "^1.1.2"
