+++ firmware/sim-recorder/Cargo.toml
@@ -3,7 +3,7 @@ resolver = "3"
 members = ["."]
 
 [package]
-name = "sim-wasm"
+name = "sim-recorder"
 version = "0.1.0"
 edition = "2024"
 
@@ -12,16 +12,11 @@ crate-type = ["cdylib"]
 
 [dependencies]
 ossm = { path = "../../ossm", features = ["sim"] }
-sim-motor = { path = "../../drivers/sim-motor" }
-sim-board = { path = "../../boards/sim-board" }
 pattern-engine = { path = "../../features/pattern-engine" }
 
 wasm-bindgen = "0.2"
-wasm-bindgen-futures = "0.4"
-embassy-time = { version = "0.5.1", features = ["wasm", "generic-queue-32"] }
+embedded-hal-async = "1.0"
 critical-section = { version = "1.2", features = ["std"] }
-log = { version = "0.4", features = ["release_max_level_info"] }
-web-sys = { version = "0.3", features = ["console"] }
 
 [profile.release]
 codegen-units = 1
