+++ firmware/sim-recorder/src/recorder.rs
